+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 version: 1.1.0
 changelog:
+
 1. DDOS BASIC (active)
 2. DDOS ALL REQUEST (active)
 3. IP REVERSE (active)
@@ -9,28 +10,39 @@ changelog:
 
 version: 1.2.0
 changelog:
+
 1. Add custom max_workers at config.json file
 2. Add more data for SPAM POST menu
 
 version: 1.2.5
 changelog:
+
 1. Add port in every request for intensive layer
 2. SMTP KILLER (active)
-2. DNS RECORD (inactive)
+3. DNS RECORD (inactive)
 
 version: 1.5.0
 changelog:
+
 1. PORT SCANNER (active)
 2. DNS RECORD (soon)
-3. 
+3.
 
 version: 1.6.0
 changelog:
+
 1. SPAM POST (update) [improvement]
 2. DataType typing union
 
 version: 1.7.0
 changelog:
+
 1. WHOIS (active)
 2. DNS RECORD (inactive)
 3. BRUTEFORCE WEBSITE (active)
+
+version: 1.8.0
+changelog:
+
+1. Add more proxy providers
+2. Inreasse total ddos at one session